The access grammar leaves the service: modules/access-model gets access-grammar, group-files, and creator
access-control.service.ts kept two very different things in one file: the
pure grammar (the YAML-subset parser, frontmatter extraction,
canonicalisation, the entry/roles/access-file parsers, and every constant
they share) and the stateful resolver + AccessControlService that loads
models from the filesystem and git. The grammar now lives in
modules/access-model/access-grammar.ts; the service, the resolver, and the
I/O-shaped loadActiveGroups/GroupsHealth stay in modules/access and import
the grammar. No compatibility re-exports — every importer, in-module and
cross-module (plugins, workflow, workspace startup, tests), points at the
grammar directly.

group-files.ts moves along with it: it is pure parsing of the same family
(parseGroupsFile/validateGroupsFile/unsafeNameReason), mergeGroupsIntoRoles
needs its GroupsIndex type, and the move dissolves the documented
"deliberate module cycle" between it and the old service file.

creator-access.ts splits the same way: Creator, CreationGrantPlan,
ICreatorAccess, and creatorPrincipal move to access-model/creator.ts, the
CreatorAccessService stays. kb-fs/locking-filesystem's type import now
points down at access-model, so kb-fs no longer imports modules/access at
all.

Three `import { type WorkspaceService }` statements (tool-helpers,
tool-manuals x2) become `import type` — they were type-only in substance
but read as runtime edges, and were the last strands of a
tool-helpers/tool-manuals/workspace runtime cycle. The module runtime-edge
graph is now acyclic.
