+++ packages/core-backend/src/modules/access-model/access-splice.ts
@@ -33,7 +33,7 @@ import {
   canonicalEmail,
   canonicalRoleName,
   accessMdDeclaresBodyRules,
-} from '../access/access-control.service.js';
+} from './access-grammar.js';
 
 /**
  * Which rule block a mutation edits.
