+++ packages/core-backend/src/modules/access/__tests__/synced-groups-writer.test.ts
@@ -1,5 +1,5 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
-import { parseGroupsFile } from '../group-files.js';
+import { parseGroupsFile } from '../../access-model/group-files.js';
 import {
   SyncedGroupsWriter,
   renderSyncedGroupsYaml,
