+++ packages/core-backend/src/modules/access/groups-admin.routes.ts
@@ -1,7 +1,7 @@
 import express from 'express';
 import type { AuthUser } from '@bevel-software/platform-shared';
 import type { IAdminAccessService } from '../admin/admin.interface.js';
-import { canonicalRoleName } from './access-control.service.js';
+import { canonicalRoleName } from '../access-model/access-grammar.js';
 import { sendError as sharedSendError, requireNonEmptyString as sharedRequireNonEmptyString } from './admin-route-helpers.js';
 import type { GroupsAdminService } from './groups-admin.service.js';
 import '../auth/auth.middleware.js'; // Express Request augmentation
