+++ packages/core-backend/src/modules/access/groups-edit.ts
@@ -20,8 +20,8 @@ import {
   EMAIL_REGEX,
   GROUP_REF_PREFIX,
   RESERVED_ROLE_NAMES,
-} from './access-control.service.js';
-import { GROUPS_YAML, parseGroupsFile, unsafeNameReason } from './group-files.js';
+} from '../access-model/access-grammar.js';
+import { GROUPS_YAML, parseGroupsFile, unsafeNameReason } from '../access-model/group-files.js';
 
 /** Bad-input failure when editing groups.yaml. */
 export class GroupsEditError extends Error {
