+++ packages/core-backend/src/modules/access/reference-scan.ts
@@ -30,7 +30,7 @@ import {
   isAccessMdPath,
   parseAccessEntry,
   stripComment,
-} from './access-control.service.js';
+} from '../access-model/access-grammar.js';
 
 /**
  * Resolve the [start, end) line range that role rewrites may touch — the YAML
