+++ packages/core-backend/src/modules/access/roles-admin.service.ts
@@ -42,8 +42,8 @@ import {
   ROLE_TOKEN_PREFIX,
   canonicalRoleName,
   canonicalEmail,
-  loadActiveGroups,
-} from './access-control.service.js';
+} from '../access-model/access-grammar.js';
+import { loadActiveGroups } from './access-control.service.js';
 import { makeRolesYamlWriteValidator } from '../access-model/roles-yaml-guard.js';
 import { renderRolesYaml } from '../access-model/render-roles-yaml.js';
 import {
@@ -57,7 +57,7 @@ import {
   RolesEditError,
   type EditResult,
 } from './roles-edit.js';
-import { GROUPS_YAML, SYNCED_GROUPS_YAML, validateGroupsFile } from './group-files.js';
+import { GROUPS_YAML, SYNCED_GROUPS_YAML, validateGroupsFile } from '../access-model/group-files.js';
 import {
   GroupsEditError,
   assertSafeGroupDisplayName,
@@ -65,7 +65,7 @@ import {
   parseGroupsModel,
 } from './groups-edit.js';
 import { capabilityRoleFor, isLegacyPeopleSetRole } from './capability-registry.js';
-import { GROUP_REF_PREFIX, RESERVED_ROLE_NAMES } from './access-control.service.js';
+import { GROUP_REF_PREFIX, RESERVED_ROLE_NAMES } from '../access-model/access-grammar.js';
 import { AdminLockedCommits } from './admin-locked-commit.js';
 import { KbReferenceScanner } from './reference-scan.js';
