+++ packages/core-backend/src/modules/access/roles-edit.ts
@@ -39,7 +39,7 @@ import {
   canonicalEmail,
   EMAIL_REGEX,
   GROUP_REF_PREFIX,
-} from './access-control.service.js';
+} from '../access-model/access-grammar.js';
 
 /** Bad-input failure when editing roles.yaml (invalid name/email, unknown role). */
 export class RolesEditError extends Error {
