+++ packages/core-backend/src/modules/access/synced-groups-committer.ts
@@ -3,7 +3,7 @@ import type { AuthUser, IWorkflowService } from '@bevel-software/platform-shared
 import type { WorkspaceService } from '../workspace/workspace.service.js';
 import { workspaceIdForBranch } from '../../shared/workspace-id.js';
 import type { AccessControlService } from './access-control.service.js';
-import { SYNCED_GROUPS_YAML } from './group-files.js';
+import { SYNCED_GROUPS_YAML } from '../access-model/group-files.js';
 import type { WorkflowEventBus } from '../workflow/event-bus.js';
 import { LockingFilesystem } from '../kb-fs/locking-filesystem.js';
 import type { SyncedGroupsWriterDeps } from './synced-groups-writer.js';
