+++ packages/core-backend/src/modules/tool-helpers/tool-context.ts
@@ -3,8 +3,8 @@ import type { IWorkflowService } from '@bevel-software/platform-shared';
 import { LockingFilesystem } from '../kb-fs/locking-filesystem.js';
 import { ReadOnlyFilesystem } from '../kb-fs/read-only-filesystem.js';
 import { makeRolesYamlWriteValidator } from '../access-model/roles-yaml-guard.js';
-import type { ICreatorAccess } from '../access/creator-access.js';
-import { type WorkspaceService } from '../workspace/workspace.service.js';
+import type { ICreatorAccess } from '../access-model/creator.js';
+import type { WorkspaceService } from '../workspace/workspace.service.js';
 import { branchForWorkspaceId } from '../../shared/workspace-id.js';
 import type { WorkflowEventBus } from '../workflow/event-bus.js';
 import type { AuthService } from '../auth/auth.service.js';
