+++ packages/core-backend/src/modules/workspace/__tests__/workspace.routes.delete.test.ts
@@ -11,7 +11,7 @@ import type { WorkflowEventBus } from '../../workflow/event-bus.js';
 import type { AuthService } from '../../auth/auth.service.js';
 import type { IAdminAccessService } from '../../admin/admin.interface.js';
 import { createWorkspaceRoutes } from '../workspace.routes.js';
-import type { ICreatorAccess } from '../../access/creator-access.js';
+import type { ICreatorAccess } from '../../access-model/creator.js';
 import type { WorkspaceService } from '../workspace.service.js';
 
 const stubCreatorAccess: ICreatorAccess = {
