+++ packages/core-backend/src/modules/workspace/__tests__/workspace.routes.download.test.ts
@@ -9,7 +9,7 @@ import type { AuthService } from '../../auth/auth.service.js';
 import { AccessConfigError } from '../../access-model/access-errors.js';
 import type { IAdminAccessService } from '../../admin/admin.interface.js';
 import { createWorkspaceRoutes } from '../workspace.routes.js';
-import type { ICreatorAccess } from '../../access/creator-access.js';
+import type { ICreatorAccess } from '../../access-model/creator.js';
 import { FolderTooLargeError, type WorkspaceService } from '../workspace.service.js';
 
 const stubCreatorAccess: ICreatorAccess = {
