+++ packages/core-backend/src/modules/workspace/workspace.routes.ts
@@ -19,7 +19,7 @@ import type { WorkspaceService } from './workspace.service.js';
 import type { AuthService } from '../auth/auth.service.js';
 import type { IAccessControl } from '../access/access-control.interface.js';
 import { resolveReadableMap, toKbRelative } from '../access-model/kb-read-filter.js';
-import type { ICreatorAccess } from '../access/creator-access.js';
+import type { ICreatorAccess } from '../access-model/creator.js';
 import { isRolesYamlPath, assertRolesYamlParsable } from '../access-model/roles-yaml-guard.js';
 import type { WorkflowEventBus } from '../workflow/event-bus.js';
 import { WorkflowDomainError } from '../../shared/domain-errors.js';
